ideater/modules/flowchart_view/requirements.txt
# Core dependencies
pydantic>=1.10.7
python-dotenv>=1.0.0

# OpenAI integration
openai>=0.27.4

# Visualization tools
graphviz>=0.20.1
pydot>=1.4.2
mermaid-cli>=0.1.1

# Utilities
numpy>=1.24.3
tqdm>=4.65.0
python-slugify>=8.0.1